requirements.txt
# Dependencias principales
Django>=5.2.0,<6.0.0
Pillow>=10.0.0

# Dependencias de producción
gunicorn>=21.2.0
whitenoise>=6.6.0
psycopg2-binary>=2.9.0

# Dependencias opcionales para validaciones avanzadas
python-magic>=0.4.27

# Dependencias de monitoreo y optimización
psutil>=5.9.0

# Dependencias de desarrollo (comentadas para producción)
# coverage>=7.3.0
# django-debug-toolbar>=4.2.0